+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 # GameVault Backend Server Changelog
 
-## Upcoming
+## 10.1.0
 
 ### Changes
 
 - Reintroduced Image Garbage Collection! Unused Images now get cleaned up from the database and the file-system. [#230](https://github.com/Phalcode/gamevault-backend/issues/230)
 - Added Configuration variables `IMAGE_GC_DISABLED` and `IMAGE_GC_INTERVAL_IN_MINUTES`
-- Server Performance Optimizations
+- Various Performance Optimizations
 
 ### Thanks